# docs/openapi/custody-v1.yaml
openapi: 3.1.0
info:
  title: Surprising Wallet Custody API
  version: 1.0.0
  description: |
    Multi-tenant custody, Console, and platform control-plane API.
    Tenant identity is derived from the authenticated credential and is never
    accepted as a trusted request field.
servers:
  - url: https://wallet.example.com
...
paths:
...
  /custody/console/v1/addresses:
    get:
      tags: [Console addresses]
      summary: List tenant addresses
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Chain'
        - name: source
          in: query
          schema: { type: string, enum: [API, CONSOLE] }
        - $ref: '#/components/parameters/Status'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Addresses
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Address' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Console addresses]
      summary: Create a tenant address from Console
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAddress' }
      responses:
        '200':
          description: Created address
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Address' }
        default: { $ref: '#/components/responses/Error' }
  /custody/console/v1/addresses/{addressId}:
    patch:
      tags: [Console addresses]
      summary: Change address label, metadata, or lifecycle status
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/AddressId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label: { type: string, maxLength: 160 }
                metadata: { type: object, additionalProperties: true }
                status: { type: string, enum: [ACTIVE, DISABLED] }
      responses:
        '200':
          description: Updated address
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Address' }
        default: { $ref: '#/components/responses/Error' }
  /custody/console/v1/assets:
    get:
      tags: [Console addresses]
      summary: List aggregate tenant assets
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Asset balances
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AssetBalance' }
        default: { $ref: '#/components/responses/Error' }
  /custody/console/v1/deposits:
    get:
      tags: [Console transfers]
      summary: List deposits
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Status'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Deposits
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Deposit' }
        default: { $ref: '#/components/responses/Error' }
  /custody/console/v1/withdrawals:
    get:
      tags: [Console transfers]
      summary: List withdrawals
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Status'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Withdrawals
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Withdrawal' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Console transfers]
      summary: Create a withdrawal from Console
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateWithdrawal' }
      responses:
        '200':
          description: Created withdrawal
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Withdrawal' }
        default: { $ref: '#/components/responses/Error' }
  /custody/console/v1/api-keys:
    get:
      tags: [Console security]
      summary: List API keys
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: API keys }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Console security]
      summary: Create an API key
      description: The plaintext secret is returned only once.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, scopes]
              properties:
                name: { type: string, maxLength: 120 }
                scopes:
                  type: array
                  uniqueItems: true
                  items:
                    type: string
                    enum:
                      - addresses:read
                      - addresses:write
                      - assets:read
                      - deposits:read
                      - withdrawals:read
                      - withdrawals:write
      responses:
        '200': { description: One-time key ID and secret }
        default: { $ref: '#/components/responses/Error' }
  /custody/console/v1/api-keys/{apiKeyId}:
    delete:
      tags: [Console security]
      summary: Revoke an API key
      security: [{ bearerAuth: [] }]
      parameters:
        - name: apiKeyId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        default: { $ref: '#/components/responses/Error' }
  /custody/console/v1/ip-allowlist:
    get:
      tags: [Console security]
      summary: Read IP allowlist state and rules
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Enforcement state and CIDR rules }
        default: { $ref: '#/components/responses/Error' }
  /custody/console/v1/ip-allowlist/enforcement:
    put:
      tags: [Console security]
      summary: Enable or disable API IP enforcement
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled: { type: boolean }
      responses:
        '200': { description: Updated enforcement state and rules }
        default: { $ref: '#/components/responses/Error' }
  /custody/console/v1/ip-allowlist/rules:
    post:
      tags: [Console security]
      summary: Add an IPv4 or IPv6 CIDR rule
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label, cidr]
              properties:
                label: { type: string, maxLength: 120 }
                cidr: { type: string, maxLength: 64 }
      responses:
        '200': { description: Created CIDR rule }
        default: { $ref: '#/components/responses/Error' }
  /custody/console/v1/ip-allowlist/rules/{ruleId}:
    delete:
      tags: [Console security]
      summary: Delete a CIDR rule
      security: [{ bearerAuth: [] }]
      parameters:
        - name: ruleId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        default: { $ref: '#/components/responses/Error' }
  /custody/console/v1/audit-log:
    get:
      tags: [Console security]
      summary: List tenant audit entries
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200': { description: Audit entries }
        default: { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: ConsoleSession
    custodyKey:
      type: apiKey
      in: header
      name: X-Custody-Key
    custodyTimestamp:
      type: apiKey
      in: header
      name: X-Custody-Timestamp
    custodyNonce:
      type: apiKey
      in: header
      name: X-Custody-Nonce
    custodySignature:
      type: apiKey
      in: header
      name: X-Custody-Signature
  parameters:
    AddressId:
      name: addressId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Chain:
      name: chain
      in: query
      schema: { type: string }
    Status:
      name: status
      in: query
      schema: { type: string }
    Search:
      name: search
      in: query
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
  responses:
    Ok:
      description: Operation completed
      content:
        application/json:
          schema:
            type: object
            required: [ok]
            properties:
              ok: { type: boolean, const: true }
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
...
    CreateAddress:
      type: object
      required: [chain]
      properties:
        chain: { type: string, pattern: '^[A-Z][A-Z0-9_]{1,31}$' }
        externalReference: { type: string, maxLength: 160 }
        label: { type: string, maxLength: 160 }
        metadata: { type: object, additionalProperties: true }
    Address:
      type: object
      required: [id, chain, network, address, metadata, source, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        chain: { type: string }
        network: { type: string }
        address: { type: string }
        memo: { type: [string, 'null'] }
        externalReference: { type: [string, 'null'] }
        label: { type: [string, 'null'] }
        metadata: { type: object, additionalProperties: true }
        source: { type: string, enum: [API, CONSOLE] }
        status: { type: string, enum: [ACTIVE, DISABLED] }
        createdAt: { type: string, format: date-time }
    AssetBalance:
      type: object
      required: [chain, assetSymbol, availableBalance, lockedBalance, totalBalance, addressCount]
      properties:
        chain: { type: string }
        assetSymbol: { type: string }
        availableBalance: { type: number }
        lockedBalance: { type: number }
        totalBalance: { type: number }
        addressCount: { type: integer, format: int64 }
    Deposit:
      type: object
      properties:
        id: { type: string, format: uuid }
        custodyAddressId: { type: string, format: uuid }
        externalReference: { type: [string, 'null'] }
        chain: { type: string }
        assetSymbol: { type: string }
        txHash: { type: string }
        logIndex: { type: integer, format: int64 }
        amount: { type: number }
        status: { type: string }
        creditedAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateWithdrawal:
      type: object
      required: [custodyAddressId, chain, assetSymbol, toAddress, amount]
      properties:
        custodyAddressId: { type: string, format: uuid }
        chain: { type: string }
        assetSymbol: { type: string }
        toAddress: { type: string, maxLength: 160 }
        amount: { type: string }
        externalReference: { type: string, maxLength: 160 }
    Withdrawal:
      type: object
      properties:
        id: { type: string, format: uuid }
        custodyAddressId: { type: string, format: uuid }
        orderNo: { type: string }
        externalReference: { type: [string, 'null'] }
        chain: { type: string }
        assetSymbol: { type: string }
        toAddress: { type: string }
        amount: { type: number }
        fee: { type: number }
        status: { type: string }
        txHash: { type: [string, 'null'] }
        errorMessage: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
...